🚧 Work in progress

+++ pages/api/clientes/index.ts
@@ -16,22 +16,25 @@ export default withApiAuthRequired(
 		if (!user.isAdmin) return res.status(403).end()
 
 		if (!req.query.path) {
-			switch (req.method) {
-				case "GET":
-					const customers = await getCustomers()
-					return res.status(200).json(customers)
-				case "POST":
-					await addCustomer(JSON.parse(req.body))
-					return res.status(201).end()
-				case "PUT":
-					const { id, values } = req.body
-					await updateCustomer(values, id)
-					return res.status(201).end()
-				case "DELETE":
-					await deleteCustomer(req.body.id)
-					return res.status(201).end()
-				default:
-					return res.status(405).end()
+			if (req.method === "GET") {
+				const customers = await getCustomers();
+				return res.status(200).json(customers);
+			}
+			else if (req.method === "POST") {
+				const id = await addCustomer(JSON.parse(req.body));
+				return res.status(201).json({ id });
+			}
+			else if (req.method === "PUT") {
+				const { id, values } = req.body;
+				await updateCustomer(values, id);
+				return res.status(201).end();
+			}
+			else if (req.method === "DELETE") {
+				await deleteCustomer(req.body.id);
+				return res.status(201).end();
+			}
+			else {
+				return res.status(405).end();
 			}
 		}
 	}
